% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_cols.R
\name{causal_cols}
\alias{causal_cols}
\alias{causal_cols<-}
\alias{set_causal_cols}
\alias{add_causal_col}
\alias{get_causal_col}
\title{Work directly with \code{causal_cols}}
\usage{
causal_cols(data)

causal_cols(data) <- value

set_causal_cols(data, ...)

add_causal_col(data, what, ..., ptype = NULL)

get_causal_col(data, what)
}
\arguments{
\item{data}{A \link{causal_tbl}.}

\item{value}{New value for \code{causal_cols}.}

\item{...}{Named attributes to add to \code{data}'s causal attributes.}

\item{what}{The causal column to get or add.}

\item{ptype}{A type to coerce a single added column to.}
}
\value{
Varies. Setter methods return the original \code{data}, perhaps invisibly.
}
\description{
These functions are aimed at developers who wish to extend \code{causal_tbl}
functionality.
}
\section{Functions}{
\itemize{
\item \code{causal_cols(data) <- value}: Set \code{causal_cols}

\item \code{set_causal_cols()}: Set column(s) for a \code{causal_col}

\item \code{add_causal_col()}: Add a single column to a \code{causal_col}

\item \code{get_causal_col()}: Get the column name of the requested variable

}}
\examples{
data <- data.frame(
  milk_first = c(0, 1, 0, 1, 1, 0, 0, 1),
  guess = c(0, 1, 0, 1, 1, 0, 0, 1)
) |>
  set_causal_cols(outcomes=guess, treatments=c(guess=milk_first))
print(data)
get_causal_col(data, "treatments")
get_causal_col(data, "outcomes")
causal_cols(data)
}
