% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{new_causal_idx}
\alias{new_causal_idx}
\alias{causal_idx}
\alias{is_causal_idx}
\alias{as_causal_idx}
\title{Construct a list of indices of type \code{causal_idx}}
\usage{
new_causal_idx(x = list())

causal_idx(x = list())

is_causal_idx(x)

as_causal_idx(x)
}
\arguments{
\item{x}{\itemize{
\item For \code{causal_idx()} and \code{new_causal_idx()}: A list of indices of type \code{causal_idx}
\item For \code{is_causal_idx()}: An object to test
\item For \code{as_causal_idx()}: An object to coerce
}}
}
\value{
A \code{causal_idx} object.
}
\description{
A \code{causal_idx} list is a list with integer vector entries which represent
indices in some other object (like a data frame). Generically it can
be used to represent a graph structure, such as an interference network
or a collection of matched objects.
The main feature of \code{causal_idx} is that the index references are preserved
through slicing and reordering. Indices that no longer refer to elements
(because of subsetting) are set to NA.
}
\section{Functions}{
\itemize{
\item \code{new_causal_idx()}: Construct a \code{causal_idx} list with minimal checks

\item \code{is_causal_idx()}: Return \code{TRUE} if an object is an \code{causal_idx} list

\item \code{as_causal_idx()}: Coerce an object to an \code{causal_idx} list

}}
\examples{
idx <- causal_idx(list(2, c(1, NA, 3), 2))
print(idx)
idx[1:2] # subsetting
idx[c(2, 1, 3)] # reordering
}
