% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_mod.R
\name{causal_mod}
\alias{causal_mod}
\alias{new_causal_mod}
\alias{is_causal_mod}
\title{Construct a fitted model column}
\usage{
causal_mod(x, idx = NULL)

new_causal_mod(x = list(), fitted = double(0), idx = seq_along(fitted))

is_causal_mod(x)
}
\arguments{
\item{x}{\itemize{
\item For \code{causal_mod()} and \code{new_causal_mod()}: A fitted model object.
For \code{causal_mod()} this should support the \code{\link[=fitted]{fitted()}} generic.
\item For \code{is_causal_mod()}: An object to test
}}

\item{idx}{A set of indices that connect observations fed into the model with
fitted values. Should contain values from 1 to the number of fitted values,
and may contain \code{NA} values for observations with no corresponding fitted
value (such as those with missing data). Defaults to a sequence over the
fitted values, with \code{NA}s determined by \code{\link[=na.action]{na.action()}}.}

\item{fitted}{A vector of fitted values.  Extracted automatically from the
model object in \code{causal_mod()}.}
}
\value{
\itemize{
\item For \code{causal_mod()} and \code{new_causal_mod()}: A \code{causal_mod} object.
\item For \code{is_causal_mod()}, a logical value.
}
}
\description{
The \code{causal_mod} class acts like a vector of fitted values, but it also
stores the fitted model for later predictions and summaries, and keeps track
of observations that have been dropped (e.g. due to missingness) or subsetted.
}
\details{
A \code{causal_mod} object can be treated and will behave as a numeric vector of
fitted values. The original model object is stored in the \code{"model"}
attribute, which can be accessed with \code{\link[=get_model_fit]{get_model_fit()}}. The \code{idx} indices
(which may have been computed automatically) are stored in the \code{"idx"}
attribute, which can be accessed with \code{\link[=get_model_idx]{get_model_idx()}}.
}
\section{Functions}{
\itemize{
\item \code{new_causal_mod()}: Construct a \code{causal_mod} with minimal checks

\item \code{is_causal_mod()}: Return \code{TRUE} if an object is a \code{causal_mod}

}}
\examples{
m <- lm(yield ~ block + N*P*K, data=npk)
causal_mod(m)

d <- rbind(NA, npk)
m_mis <- lm(yield ~ block + N*P*K, data=d)
fitted(m_mis) # length doesn't match rows of `d`
causal_mod(m_mis) # NA for missing value
attr(causal_mod(m_mis), "idx")
attr(causal_mod(m_mis), "mod")

}
