% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_tbl.R
\name{new_causal_tbl}
\alias{new_causal_tbl}
\alias{causal_tbl}
\alias{as_causal_tbl}
\alias{is_causal_tbl}
\title{Build a causal data frame}
\usage{
new_causal_tbl(..., .outcome = NULL, .treatment = NULL)

causal_tbl(..., .outcome = NULL, .treatment = NULL)

as_causal_tbl(x)

is_causal_tbl(x)
}
\arguments{
\item{...}{passed on to \code{\link[=tibble]{tibble()}}}

\item{.outcome}{the column containing the outcome variable (tidy-selected).
Can be set later with \code{\link[=set_outcome]{set_outcome()}}.}

\item{.treatment}{the column containing the treatment variable (tidy-selected).
Can be set later with \code{\link[=set_treatment]{set_treatment()}}.}

\item{x}{A data frame to be checked or coerced}
}
\value{
A \code{causal_tbl} object
}
\description{
A \code{causal_tbl} is a tibble with additional attribute information stored
in \link{causal_cols}.  See the 'Internal structure' for more on the structure of this
attribute.
}
\details{
At its core, a \code{causal_tbl} is just a tibble, and it should behave like
a tibble in every meaningful way.
What sets a \code{causal_tbl} apart is that it keeps track of \emph{causal columns}:
variables or objects which play a particular causal role.
These can be accessed with the various getter and setter functions included
in this package, like \code{\link[=get_outcome]{get_outcome()}} and \code{\link[=set_outcome]{set_outcome()}}.
}
\section{Functions}{
\itemize{
\item \code{new_causal_tbl()}: Construct a \code{causal_tbl} with no checks

\item \code{as_causal_tbl()}: Coerce a data frame to a \code{causal_tbl}

\item \code{is_causal_tbl()}: Return \code{TRUE} if a data frame is a \code{causal_tbl}

}}
\section{Internal structure}{
The \link{causal_cols} attribute is considered mostly internal, and end users
do not have to worry about its internal structure. However, for those
developing packages based off of \code{causal_tbl}, it is useful to understand the
underlying structure of \code{causal_cols}.

The \code{causal_cols} attribute is a named list, with each element corresponding
to a type of causal variable or object: \code{outcomes}, \code{treatments}, \code{panel_unit}, but
also potentially \code{pscore}, \code{matches}, \code{model}, etc.
Each of these elements is a character vector, with each element being a name
of a column in the data frame.
For some variables, this vector should be of length 1, but for other
variables, there may be multiple columns of that type.
So, for example, if a package author was developing methods for causal
inference with multiple continuous treatments, the \code{treatment} element
of \code{causal_cols} could have an entry for each treatment column.

The optional \code{\link[=names]{names()}} of the columns within a particular element of
\code{causal_cols} convey information on any associated variable.  For example,
the treatment variable is by default associated with a particular outcome.
And a propensity score or outcome model is associated with a particular
treatment or outcome variable.

The column names stored within any part of \code{causal_cols} will be automatically
updated if columns are renamed, or set to \code{NULL} if columns are dropped.
This reassignment happens automatically and silently in all cases.
It is the responsibility of implementers of particular methods to check
that a \code{causal_tbl} has the necessary columns set via helpers like
\code{\link[=has_treatment]{has_treatment()}}, \code{\link[=has_outcome]{has_outcome()}}, etc.
}

\examples{
data <- causal_tbl(
  milk_first = c(0, 1, 0, 1, 1, 0, 0, 1),
  guess = c(0, 1, 0, 1, 1, 0, 0, 1)
)
is_causal_tbl(data)
print(data)

}
