% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causaltbl-package.R
\docType{package}
\name{causaltbl-package}
\alias{causaltbl}
\alias{causaltbl-package}
\title{causaltbl: Tidy Causal Data Frames and Tools}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a 'causal_tbl' class for causal inference. A 'causal_tbl' keeps track of information on the roles of variables like treatment and outcome, and provides functionality to store models and their fitted values as columns in a data frame.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CoryMcCartan/causaltbl}
  \item \url{http://corymccartan.com/causaltbl/}
  \item Report bugs at \url{https://github.com/CoryMcCartan/causaltbl/issues}
}

}
\author{
\strong{Maintainer}: Cory McCartan \email{cmccartan@g.harvard.edu}

Authors:
\itemize{
  \item Christopher T. Kenny \email{christopherkenny@fas.harvard.edu}
}

}
\keyword{internal}
