% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_mod.R
\name{get_model_fit}
\alias{get_model_fit}
\alias{get_model_idx}
\title{Extract the model object and observation indices from a \code{causal_mod}}
\usage{
get_model_fit(x)

get_model_idx(x)
}
\arguments{
\item{x}{The \code{causal_mod} object}
}
\value{
\itemize{
\item For \code{get_model_fit()}: A fitted model object
\item For \code{get_model_idx()}: An integer vector of indices
}
}
\description{
Extract the model object and observation indices from a \code{causal_mod}
}
\examples{
m <- lm(yield ~ block + N*P*K, data=npk)
x <- causal_mod(m)
identical(get_model_fit(x), m) # TRUE
get_model_idx(x)

}
