% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{new_indices}
\alias{new_indices}
\alias{indices}
\alias{is_indices}
\alias{as_indices}
\title{Construct a list of indices}
\usage{
new_indices(x = list())

indices(x = list())

is_indices(x)

as_indices(x)
}
\arguments{
\item{x}{\itemize{
\item For \code{indices()} and \code{new_indices()}: A list of indices
\item For \code{is_indices()}: An object to test
\item For \code{as_indices()}: An object to coerce
}}
}
\value{
An \code{indices} object.
}
\description{
An \code{indices} list is a list with integer vector entries which represent
indices in some other object (like a data frame). Generically it can
be used to represent a graph structure, such as an interference network
or a collection of matched objects.
The main feature of \code{indices} is that the index references are preserved
through slicing and reordering. Indices that no longer refer to elements
(because of subsetting) are set to NA.
}
\section{Functions}{
\itemize{
\item \code{new_indices()}: Construct an \code{indices} list with minimal checks

\item \code{is_indices()}: Return \code{TRUE} if an object is an \code{indices} list

\item \code{as_indices()}: Coerce an object to an \code{indices} list

}}
\examples{
idx <- indices(list(2, c(1, NA, 3), 2))
print(idx)
idx[1:2] # subsetting
idx[c(2, 1, 3)] # reordering
}
