% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_cols.R
\name{set_causal_col}
\alias{set_causal_col}
\alias{get_causal_col}
\alias{add_causal_col}
\title{Define Custom Causal Attributes}
\usage{
set_causal_col(data, ...)

get_causal_col(data, what)

add_causal_col(data, ...)
}
\arguments{
\item{data}{a data frame or \code{causal_tbl}}

\item{...}{named attributes to add to \code{data}'s causal attributes}

\item{what}{the causal attribute to get}
}
\value{
A \code{causal_tbl}

For \code{get_causal_col()} the column name of the requested variable

For \code{add_causal_col()} A \code{causal_tbl}
}
\description{
Define Custom Causal Attributes
}
\examples{
data <- data.frame(
  milk_first = c(0, 1, 0, 1, 1, 0, 0, 1),
  guess = c(0, 1, 0, 1, 1, 0, 0, 1)
) |>
  set_causal_col(milk_first, "treatment")
print(data) # a causal_tbl
get_causal_col(data, "treatment")
}
