% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_cols.R
\name{set_outcome}
\alias{set_outcome}
\alias{get_outcome}
\alias{has_outcome}
\alias{pull_outcome}
\title{Define an outcome variable for a \code{causal_tbl}}
\usage{
set_outcome(data, outcome)

get_outcome(data)

has_outcome(data)

pull_outcome(data)
}
\arguments{
\item{data}{a data frame or \code{causal_tbl}}

\item{outcome}{the column containing the outcome variable (tidy-selected).
Must be numeric or coercible to numeric.}
}
\value{
A \code{causal_tbl}

For \code{get_outcome()} the column name of the outcome variable

For \code{has_outcome()}, \code{TRUE} if there is an outcome variable set

For \code{pull_outcome()}, the vector of the outcome variable.
}
\description{
Define an outcome variable for a \code{causal_tbl}
}
\examples{
data <- data.frame(
  milk_first = c(0, 1, 0, 1, 1, 0, 0, 1),
  guess = c(0, 1, 0, 1, 1, 0, 0, 1)
) |>
  set_outcome(guess)
print(data) # a causal_tbl
get_outcome(data)
}
