% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_cols.R
\name{set_treatment}
\alias{set_treatment}
\alias{get_treatment}
\alias{has_treatment}
\alias{get_panel}
\alias{has_panel}
\title{Define a treatment variable for a \code{causal_tbl}}
\usage{
set_treatment(data, treatment)

get_treatment(data)

has_treatment(data)

get_panel(data)

has_panel(data)
}
\arguments{
\item{data}{a data frame or \code{causal_tbl}}

\item{outcome}{the column containing the treatment variable (tidy-selected).
Must be numeric or coercible to numeric.}
}
\value{
A \code{causal_tbl}

For \code{get_treatment()} the column name of the treatment variable

For \code{has_treatment()}, \code{TRUE} if there is a treatment variable set

For \code{get_panel()} a list with the column names of the unit and time variables

For \code{has_panel()}, \code{TRUE} if there is panel data structure
}
\description{
Define a treatment variable for a \code{causal_tbl}
}
\examples{
data <- data.frame(
  milk_first = c(0, 1, 0, 1, 1, 0, 0, 1),
  guess = c(0, 1, 0, 1, 1, 0, 0, 1)
) |>
  set_treatment(milk_first)
print(data) # a causal_tbl
get_treatment(data)
}
