% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_tbl.R
\name{new_causal_tbl}
\alias{new_causal_tbl}
\alias{causal_tbl}
\alias{as_causal_tbl}
\title{Build a causal data frame}
\usage{
new_causal_tbl(..., .outcome = NULL, .treatment = NULL)

causal_tbl(..., .outcome = NULL, .treatment = NULL)

as_causal_tbl(x)
}
\arguments{
\item{...}{passed on to \code{\link[=tibble]{tibble()}}}

\item{.outcome}{the column containing the outcome variable (tidy-selected).
Can be set later with \code{\link[=set_outcome]{set_outcome()}}.}

\item{.treatment}{the column containing the treatment variable (tidy-selected).
Can be set later with \code{\link[=set_treatment]{set_treatment()}}.}

\item{x}{A data frame to be coerced}
}
\value{
A \code{causal_tbl} object
}
\description{
A \code{causal_tbl} is a tibble with additional attribute information stored
in \code{causal_cols}.  See the 'Details' for more on the structure of this
attribute.
}
\section{Functions}{
\itemize{
\item \code{new_causal_tbl()}: Construct a \code{causal_tbl} with no checks

\item \code{as_causal_tbl()}: Coerce a data frame to a \code{causal_tbl}

}}
\examples{
causal_tbl(
  milk_first = c(0, 1, 0, 1, 1, 0, 0, 1),
  guess = c(0, 1, 0, 1, 1, 0, 0, 1)
)

}
