% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_cols.R
\name{set_outcome}
\alias{set_outcome}
\alias{get_outcome}
\title{Use \code{outcome} Attribute for a \code{causal_tbl}}
\usage{
set_outcome(x, .outcome)

get_outcome(x)
}
\arguments{
\item{x}{a \code{causal_tbl}}

\item{.outcome}{column name of outcome as a string.}
}
\value{
for \code{set_outcome()}, a \code{causal_tbl}, otherwise the column name of the outcome
}
\description{
Use \code{outcome} Attribute for a \code{causal_tbl}
}
\examples{
causal_tbl(
  milk_first = c(0, 1, 0, 1, 1, 0, 0, 1),
  guess = c(0, 1, 0, 1, 1, 0, 0, 1),
  correct = c(1, 1, 1, 1, 1, 1, 1, 1)
) |>
set_outcome("correct")
}
