% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_cols.R
\name{set_outcome}
\alias{set_outcome}
\alias{get_outcome}
\title{Use \code{outcome} Attribute for a \code{causal_tbl}}
\usage{
set_outcome(data, outcome)

get_outcome(data)
}
\arguments{
\item{data}{a data frame or \code{causal_tbl}}

\item{outcome}{column name of outcome (tidy-selected)}
}
\value{
A \code{causal_tbl}, otherwise the column name of the outcome

For \code{get_outcome()} the column name of the outcome variable
}
\description{
Use \code{outcome} Attribute for a \code{causal_tbl}
}
\examples{
data <- data.frame(
  milk_first = c(0, 1, 0, 1, 1, 0, 0, 1),
  guess = c(0, 1, 0, 1, 1, 0, 0, 1)
) |>
  set_outcome(guess)
print(data) # a causal_tbl
get_outcome(data)
}
