% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_cols.R
\name{set_panel}
\alias{set_panel}
\alias{get_panel}
\alias{has_panel}
\title{Define a panel data structure for a \code{causal_tbl}}
\usage{
set_panel(data, unit, time)

get_panel(data)

has_panel(data)
}
\arguments{
\item{data}{a data frame or \code{causal_tbl}}

\item{unit}{the column indexing treatment units (tidy-selected)}

\item{time}{the column indexing treatment time (tidy-selected)}
}
\value{
A \code{causal_tbl}

For \code{get_panel()} a list with the column names of the unit and time variables

For \code{has_panel()}, \code{TRUE} if there is panel data structure
}
\description{
Define a panel data structure for a \code{causal_tbl}
}
\examples{
data <- data.frame(
  id = c("a", "a", "a", "a", "b", "b", "b", "b"),
  year = rep(2015:2018, 2),
  trt = c(0, 0, 0, 0, 0, 0, 1, 1),
  y = c(1, 3, 2, 3, 2, 4, 4, 5)
) |>
  set_panel(unit=id, time=year)
print(data) # a causal_tbl
get_panel(data)
}
