% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_cols.R
\name{set_treatment}
\alias{set_treatment}
\alias{get_treatment}
\title{Use \code{treatment} Attribute for a \code{causal_tbl}}
\usage{
set_treatment(data, treatment)

get_treatment(data)
}
\arguments{
\item{data}{a data frame or \code{causal_tbl}}

\item{treatment}{column name of treatment (tidy-selected)}
}
\value{
A \code{causal_tbl}

For \code{get_treatment()} the column name of the treatment variable
}
\description{
Use \code{treatment} Attribute for a \code{causal_tbl}
}
\examples{
data <- data.frame(
  milk_first = c(0, 1, 0, 1, 1, 0, 0, 1),
  guess = c(0, 1, 0, 1, 1, 0, 0, 1)
) |>
  set_treatment(milk_first)
print(data) # a causal_tbl
get_treatment(data)
}
