% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_cols.R
\name{set_treatment}
\alias{set_treatment}
\alias{get_treatment}
\title{Use \code{treatment} Attribute for a \code{causal_tbl}}
\usage{
set_treatment(x, .treatment)

get_treatment(x)
}
\arguments{
\item{x}{a \code{causal_tbl}}

\item{.treatment}{column name of treatment as a string.}
}
\value{
for \code{set_treatment()}, a \code{causal_tbl}, otherwise the column name of the treatment
}
\description{
Use \code{treatment} Attribute for a \code{causal_tbl}
}
\examples{
causal_tbl(
  milk_first = c(0, 1, 0, 1, 1, 0, 0, 1),
  guess = c(0, 1, 0, 1, 1, 0, 0, 1)
) |>
set_treatment("milk_first")
}
